refactor: like 파일 이동 및 불필요한 파일 삭제
