--- src/main/java/com/est/runtime/post/dto/PostDto.java
@@ -1,4 +1,0 @@
-package com.est.runtime.post.dto;
-
-public class PostDto {
-}
